feat: add script to imprint frame number and time on a video
This commit adds a new script that uses FFmpeg to imprint the frame number and time on a video. The script takes in a video file path and an output file path as command line arguments. It uses the `drawtext` filter of FFmpeg to add the frame number and time as text on each frame of the video. The resulting video is saved to the specified output file path. The script also includes options for verbose mode and dry run.